# swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Collecto API
  description: API de autenticación y gestión de sesión para la plataforma Collecto
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/collecto/Collecto/1.0.0
  - url: http://localhost:3000
    description: Servidor de desarrollo local
paths:
...
  /api/orders:
    post:
      tags:
        - Ordenes
      summary: Crear una nueva orden
      description: Permite crear una nueva orden de compra o venta
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                advertId:
                  type: string
                  description: ID del anuncio que se quiere comprar o vender
      responses:
        '201':
          description: Orden creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Orden creada"
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: El anuncio no está disponible para compra
        '404':
          description: Anuncio no encontrado
  /api/orders/{id}:
    get:
      tags:
        - Ordenes
      summary: Obtener detalles de una orden
      description: Permite obtener los detalles de una orden mediante su ID
      operationId: getOrderDetails
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la orden
          schema:
            type: string
      responses:
        '200':
          description: Detalles de la orden
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '404':
          description: Orden no encontrada
  /api/orders/{id}/status:
    patch:
      tags:
        - Ordenes
      summary: Actualizar estado de la orden
      description: Permite actualizar el estado de una orden (ej. pendiente, pagado, enviado)
      operationId: updateOrderStatus
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la orden
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: ['pending', 'paid', 'shipped', 'delivered', 'cancelled']
                  description: Nuevo estado de la orden
      responses:
        '200':
          description: Estado de la orden actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Estado de la orden actualizado"
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Estado inválido
        '404':
          description: Orden no encontrada
        '403':
          description: No tienes permiso para actualizar el estado de esta orden
  /api/ordres/{id}:
    delete:
      tags:
        - Ordenes
      summary: Cancelar una orden
      description: Permite cancelar una orden existente
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la orden
          schema:
            type: string
      responses:
        '200':
          description: Orden cancelada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Orden cancelada"
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Solo se pueden cancelar las órdenes en estado "pending"
        '404':
          description: Orden no encontrada
        '403':
          description: No tienes permiso para cancelar esta orden
  /api/orders/me:
    get:
      tags:
        - Ordenes
      summary: Obtener todas las órdenes del usuario autenticado
      description: Devuelve todas las órdenes de compra y venta asociadas al usuario autenticado
      operationId: getAllUserOrders
      responses:
        '200':
          description: Lista de órdenes del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: No tienes órdenes
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        buyerId:
          type: string
          description: ID del comprador
        sellerId:
          type: string
          description: ID del vendedor
        advertId:
          type: string
          description: ID del anuncio relacionado
        price:
          type: number
          format: double
          description: Precio del pedido
        commission:
          type: number
          format: double
          description: Comisión asociada al pedido
        paymentStatus:
          type: string
          enum: ['pending', 'paid', 'shipped', 'delivered', 'cancelled']
          description: Estado del pago
        paymentID:
          type: string
          description: ID del pago asociado
        shippingMethod:
          type: string
          description: Método de envío
        shippingAddress:
          type: string
          description: Dirección de envío
        trackingCode:
          type: string
          description: Código de seguimiento
        notes:
          type: string
          description: Notas adicionales
        createdAt:
          type: string
          format: date-time
          description: Fecha de creación del pedido
        updatedAt:
          type: string
          format: date-time
          description: Fecha de la última actualización del pedido
...